Accept speech recognition results wiht both ß and ss
This change fixes the problem reported in
https://community.smartergerman.com/c/post-any-question-here/a2-42-assisted-preaching-problems-with-entering-ss
affecting those of us living in Switzerland and using Chrome.

I updated only the versions of these files that were most recently
updated & are actually used in the course. There are different versions
of script.js under v1/ and v2/ subdirectories as well, but they seem to
be unused & contain older versions of the tool, so I did not update
them.

I tested locally that the fix works for me by replacing script.js with
my version using the browser's Dev Tools.